tests made using ai
