+++ tests/conftest.py
@@ -1,3 +1,4 @@
+import os
 import shutil
 import tempfile
 from pathlib import Path
@@ -5,9 +6,22 @@ from typing import Any, Generator
 
 import pytest
 
+from haddock.libs.libcapri import CAPRI
 from haddock.libs.libontology import PDBFile
 
 from . import golden_data
+
+
+def remove_aln_files(class_name):
+    """Helper function to remove intermediary alignment files."""
+    file_l = [
+        Path(class_name.path, "blosum62.izone"),
+        Path(class_name.path, "blosum62_A.aln"),
+        Path(class_name.path, "blosum62_B.aln"),
+    ]
+    for f in file_l:
+        if f.exists():
+            os.unlink(f)
 
 
 @pytest.fixture(name="protprot_input_list")
@@ -215,3 +229,264 @@ def fixture_protprot_onechain_list() -> Generator[list[PDBFile], Any, Any]:
         pdb_obj_2 = PDBFile(file_name=dst_prot_2, path=Path(dst_prot_2).parent)
 
         yield [pdb_obj_1, pdb_obj_2]
+
+
+# ---------------------------------------------------------------------------
+# CAPRI object params fixtures
+# ---------------------------------------------------------------------------
+
+
+@pytest.fixture(name="params")
+def fixture_params():
+    """???"""
+    return {
+        "receptor_chain": "A",
+        "ligand_chains": ["B"],
+        "aln_method": "sequence",
+        "allatoms": False,
+        "keep_hetatm": False,
+    }
+
+
+@pytest.fixture(name="params_all")
+def fixture_params_all():
+    """???"""
+    return {
+        "receptor_chain": "A",
+        "ligand_chains": ["B"],
+        "aln_method": "sequence",
+        "allatoms": True,
+        "keep_hetatm": False,
+    }
+
+
+@pytest.fixture(name="params_hetatm")
+def fixture_params_hetatm():
+    """Parameters when HETATM must be kept."""
+    return {
+        "receptor_chain": "A",
+        "ligand_chains": ["B"],
+        "aln_method": "sequence",
+        "allatoms": False,
+        "keep_hetatm": True,
+    }
+
+
+# ---------------------------------------------------------------------------
+# CAPRI object fixtures
+# ---------------------------------------------------------------------------
+
+
+@pytest.fixture(name="protdna_caprimodule")
+def fixture_protdna_caprimodule(protdna_input_list, params):
+    """Protein-DNA CAPRI module."""
+    reference = protdna_input_list[0]
+    model = protdna_input_list[1]
+    _path = protdna_input_list[0].path
+    capri = CAPRI(
+        identificator=42,
+        reference=reference,
+        model=model,
+        path=_path,
+        params=params,
+    )
+
+    yield capri
+
+    remove_aln_files(capri)
+
+
+@pytest.fixture(name="protlig_caprimodule")
+def fixture_protlig_caprimodule(protlig_input_list, params):
+    """Protein-Ligand CAPRI module."""
+    reference = protlig_input_list[0]
+    model = protlig_input_list[1]
+    _path = protlig_input_list[0].path
+    capri = CAPRI(
+        identificator=42,
+        reference=reference,
+        model=model,
+        path=_path,
+        params=params,
+    )
+
+    yield capri
+
+    remove_aln_files(capri)
+
+
+@pytest.fixture(name="protprot_caprimodule")
+def fixture_protprot_caprimodule(protprot_input_list, params):
+    """Protein-Protein CAPRI module."""
+    reference = protprot_input_list[0]
+    model = protprot_input_list[1]
+    _path = protprot_input_list[0].path
+    capri = CAPRI(
+        identificator=str(42),
+        reference=reference,
+        model=model,
+        path=_path,
+        params=params,
+    )
+
+    yield capri
+
+    remove_aln_files(capri)
+
+
+@pytest.fixture(name="protprot_allatm_caprimodule")
+def fixture_protprot_allatm_caprimodule(protprot_input_list, params_all):
+    """Protein-Protein CAPRI module."""
+    reference = protprot_input_list[0]
+    model = protprot_input_list[1]
+    _path = protprot_input_list[0].path
+    capri = CAPRI(
+        identificator=42,
+        reference=reference,
+        model=model,
+        path=_path,
+        params=params_all,
+    )
+
+    yield capri
+
+    remove_aln_files(capri)
+
+
+@pytest.fixture(name="protprot_1bkd_caprimodule")
+def fixture_protprot_1bkd_caprimodule(protprot_1bkd_input_list, params):
+    """Protein-Protein CAPRI module for target 1BKD."""
+    reference = protprot_1bkd_input_list[0]
+    model = protprot_1bkd_input_list[1]
+    _path = protprot_1bkd_input_list[0].path
+    capri = CAPRI(
+        identificator=str(42),
+        reference=reference,
+        model=model,
+        path=_path,
+        params=params,
+    )
+
+    yield capri
+
+    remove_aln_files(capri)
+
+
+@pytest.fixture(name="protprot_onechain_ref_caprimodule")
+def fixture_protprot_onechain_ref_caprimodule(protprot_onechain_list, params):
+    """Protein-Protein CAPRI module with a single chain structure as ref."""
+
+    # reference, model = protprot_onechain_list
+    model, reference = protprot_onechain_list
+    _path = reference.path
+
+    capri = CAPRI(
+        identificator=str(42),
+        reference=reference,
+        model=model,
+        path=_path,
+        params=params,
+    )
+
+    yield capri
+
+    remove_aln_files(capri)
+
+
+@pytest.fixture(name="protprot_onechain_mod_caprimodule")
+def fixture_protprot_onechain_mod_caprimodule(protprot_onechain_list, params):
+    """Protein-Protein CAPRI module with a single chain structure as model."""
+
+    reference, model = protprot_onechain_list
+    _path = reference.path
+
+    capri = CAPRI(
+        identificator=str(42),
+        reference=reference,
+        model=model,
+        path=_path,
+        params=params,
+    )
+
+    yield capri
+
+    remove_aln_files(capri)
+
+
+@pytest.fixture(name="protdna_caprimodule_cg")
+def fixture_protdna_caprimodule_cg(protdna_input_list_cg, params):
+    """Protein-DNA CAPRI module (coarse-grained)."""
+    reference = protdna_input_list_cg[0]
+    model = protdna_input_list_cg[1]
+    _path = protdna_input_list_cg[0].path
+    capri = CAPRI(
+        identificator=42,
+        reference=reference,
+        model=model,
+        path=_path,
+        params=params,
+        ff="martini2",
+    )
+
+    yield capri
+
+    remove_aln_files(capri)
+
+
+@pytest.fixture(name="protprot_caprimodule_cg")
+def fixture_protprot_caprimodule_cg(protprot_input_list_cg, params):
+    """Protein-Protein CAPRI module (coarse-grained)."""
+    reference = protprot_input_list_cg[0]
+    model = protprot_input_list_cg[1]
+    _path = protprot_input_list_cg[0].path
+    capri = CAPRI(
+        identificator=str(42),
+        reference=reference,
+        model=model,
+        path=_path,
+        params=params,
+        ff="martini2",
+    )
+
+    yield capri
+
+    remove_aln_files(capri)
+
+
+@pytest.fixture(name="protprot_allatm_caprimodule_cg")
+def fixture_protprot_allatm_caprimodule_cg(protprot_input_list_cg, params_all):
+    """Protein-Protein CAPRI module, all atoms (coarse-grained)."""
+    reference = protprot_input_list_cg[0]
+    model = protprot_input_list_cg[1]
+    _path = protprot_input_list_cg[0].path
+    capri = CAPRI(
+        identificator=42,
+        reference=reference,
+        model=model,
+        path=_path,
+        params=params_all,
+        ff="martini2",
+    )
+
+    yield capri
+
+    remove_aln_files(capri)
+
+
+@pytest.fixture(name="prot_hetatmlig_caprimodule")
+def fixture_prot_hetatmlig_caprimodule(prot_HETATMlig_input_list, params_hetatm):
+    """Protein-Ligand CAPRI module (HETATM ligand)."""
+    reference = prot_HETATMlig_input_list[0]
+    model = prot_HETATMlig_input_list[1]
+    _path = prot_HETATMlig_input_list[1].path
+    capri = CAPRI(
+        identificator=42,
+        reference=reference,
+        model=model,
+        path=_path,
+        params=params_hetatm,
+    )
+
+    yield capri
+
+    remove_aln_files(capri)
